parse_linux: avoid commit check from old kernel version
For example if BSP was based on v6.3.24, it don't need to check
v6.1, v6.2 kernel, because these are already included on BSP.
It will happen offten if this script is used more longer term.
This patch avoid it.

One note is that we want to avoid new version case too in the future.
For example in above BSP case, we don't think BSP backports patch
from v8.17 kernel, etc. It is T.B.D for now.

Signed-off-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>